[Map] create component for zoom listener
